backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
social-auth-app-django==5.4.0
Pillow==11.0.0
reportlab==4.0.4
qrcode==7.4.2
django-cors-headers==4.3.1
python-dotenv==1.0.0
google-auth==2.29.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
vercel==0.3.4
whitenoise==6.6.0
gunicorn==20.1.0
razorpay==1.4.1

# Machine Learning Dependencies
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2